Added product model and changed price to float for all dtos

+++ ThAmCo-Products-WebService/Dtos/InsertProductDto.cs
@@ -4,7 +4,7 @@
     {
         public string ProductName { get; set; } = string.Empty;
         public int Quantity { get; set; }
-        public decimal Price { get; set; }
+        public float Price { get; set; }
         public string Description { get; set; } = string.Empty;
     }
 }
